+++ src/event/__tests__/index.test.js
@@ -3,7 +3,7 @@
 import {from} from 'most'
 import {createEvent, type Event} from '..'
 
-import {getSpyCalls, spy} from 'effector/fixtures/test-utils'
+import {getSpyCalls, spy} from 'effector/fixtures'
 import {show} from 'effector/datatype/step/show'
 
 describe('symbol-observable support', () => {