+++ src/event/createEvent.js
@@ -4,5 +4,5 @@ import {eventFabric} from './eventFabric'
 import type {Event} from './index.h'
 
 export function createEvent<Payload>(name?: string): Event<Payload> {
- return eventFabric({name, domainName: ''})
+ return eventFabric({name})
 }