+++ src/fixtures/index.js
@@ -1,0 +1,3 @@
+//@flow
+
+export {delay, spy, getSpyCalls} from './testUtils'